requirements.txt
selenium>=2.53.6
pytest>=2.9.2
pytest-allure-adaptor>=1.7.5